* Fix bug with y position appearing differently than expected on right size of Beam

+++ include/AllegroFlare/MusicNotation/Beam.hpp
@@ -24,10 +24,10 @@ namespace AllegroFlare
       private:
          float staff_line_distance;
          float start_x;
-         float start_staff_pos;
+         int start_staff_pos;
          Beam::Alignment start_alignment;
          float end_x;
-         float end_staff_pos;
+         int end_staff_pos;
          Beam::Alignment end_alignment;
          ALLEGRO_COLOR color;
          std::vector<std::tuple<int, float, float>> secondary_beams;
@@ -37,25 +37,25 @@ namespace AllegroFlare
 
 
       public:
-         Beam(float staff_line_distance=0, float start_x=0, float start_staff_pos=0, Beam::Alignment start_alignment=Beam::Alignment::UNDEFINED, float end_x=0.0f, float end_staff_pos=0.0f, Beam::Alignment end_alignment=Beam::Alignment::UNDEFINED, ALLEGRO_COLOR color=ALLEGRO_COLOR{1, 1, 1, 1}, std::vector<std::tuple<int, float, float>> secondary_beams={});
+         Beam(float staff_line_distance=0, float start_x=0, int start_staff_pos=0, Beam::Alignment start_alignment=Beam::Alignment::UNDEFINED, float end_x=0.0f, int end_staff_pos=0, Beam::Alignment end_alignment=Beam::Alignment::UNDEFINED, ALLEGRO_COLOR color=ALLEGRO_COLOR{1, 1, 1, 1}, std::vector<std::tuple<int, float, float>> secondary_beams={});
          ~Beam();
 
          void set_staff_line_distance(float staff_line_distance);
          void set_start_x(float start_x);
-         void set_start_staff_pos(float start_staff_pos);
+         void set_start_staff_pos(int start_staff_pos);
          void set_start_alignment(Beam::Alignment start_alignment);
          void set_end_x(float end_x);
-         void set_end_staff_pos(float end_staff_pos);
+         void set_end_staff_pos(int end_staff_pos);
          void set_end_alignment(Beam::Alignment end_alignment);
          void set_color(ALLEGRO_COLOR color);
          void set_secondary_beams(std::vector<std::tuple<int, float, float>> secondary_beams);
          void set_render_with_debugging_visuals(bool render_with_debugging_visuals);
          float get_staff_line_distance() const;
          float get_start_x() const;
-         float get_start_staff_pos() const;
+         int get_start_staff_pos() const;
          Beam::Alignment get_start_alignment() const;
          float get_end_x() const;
-         float get_end_staff_pos() const;
+         int get_end_staff_pos() const;
          Beam::Alignment get_end_alignment() const;
          ALLEGRO_COLOR get_color() const;
          std::vector<std::tuple<int, float, float>> get_secondary_beams() const;
